+++ pytorch/feedforward.py
@@ -165,11 +165,11 @@ class FeedForward(nn.Module):
 
 if __name__ == "__main__":
 	trainset = torchvision.datasets.MNIST('./data', transform=None, download=True, train=True)
-	X_train = trainset.data.numpy().astype(np.float32) / 255
+	X_train = trainset.data.numpy() / 255
 	y_train = trainset.targets.numpy()
 	
 	testset = torchvision.datasets.MNIST('./data', transform=None, download=True, train=False)
-	X_test = testset.data.numpy().astype(np.float32) / 255
+	X_test = testset.data.numpy() / 255
 	y_test = testset.targets.numpy()
 	
 	model = FeedForward()
